gpt/backend/requirements.txt
python-dotenv
uvicorn
fastapi
sqlmodel
pydantic[email]
google.generativeai
bcrypt
flask
flask_bcrypt
flask_cors
stripe
langchain_community
langchain
pymongo
langchain_openai
twilio
unstructured
networkx
pandas
openpyxl
chromadb